Backend/tests/test_critical_security.py
"""
Critical Security and Business Logic Tests

Tests for the most critical security vulnerabilities and business logic paths.
These tests ensure the application is secure and functions correctly.
"""

import pytest
from unittest.mock import Mock, patch
from Backend.src.services.auth_service import AuthService
from Backend.src.utils.error_handling import ServiceError, ValidationError, AuthenticationError


class TestCriticalSecurity:
    """Test critical security vulnerabilities and fixes"""

    def test_sql_injection_protection(self):
        """Test that SQL injection attempts are properly sanitized"""
        from Backend.src.utils.sql_injection_protection import sanitize_sql_input

        # Test malicious inputs
        malicious_inputs = [
            "'; DROP TABLE users; --",
            "1' OR '1'='1",
            "<script>alert('xss')</script>",
            "javascript:alert('xss')",
            "../../../../etc/passwd"
        ]

        for malicious_input in malicious_inputs:
            sanitized = sanitize_sql_input(malicious_input)
            # Ensure dangerous characters are escaped or removed
            assert "'" not in sanitized or sanitized.count("'") < malicious_input.count("'")
            assert "<script>" not in sanitized

    def test_xss_protection(self):
        """Test XSS protection in user inputs"""
        from Backend.src.utils.input_sanitization import sanitize_html

        xss_payloads = [
            "<script>alert('xss')</script>",
            "<img src=x onerror=alert('xss')>",
            "javascript:alert('xss')",
            "<iframe src='javascript:alert(\"xss\")'></iframe>"
        ]

        for payload in xss_payloads:
            sanitized = sanitize_html(payload)
            assert "<script>" not in sanitized
            assert "javascript:" not in sanitized
            assert "onerror" not in sanitized

    @patch('Backend.src.services.auth_service.firebase_auth')
    def test_brute_force_protection(self, mock_auth):
        """Test account lockout after failed login attempts"""
        # Mock Firebase auth to always fail
        mock_auth.verify_id_token.side_effect = Exception("Invalid token")

        # Create a valid JWT token format with email for testing
        import jwt
        import os
        test_email = "test@example.com"
        token_data = {"email": test_email, "exp": 9999999999}
        test_token = jwt.encode(token_data, "test_key", algorithm="HS256")

        # Simulate multiple failed login attempts
        for i in range(6):  # More than MAX_FAILED_LOGIN_ATTEMPTS
            result = AuthService.login_with_id_token(test_token)
            if i < 5:  # First 5 should fail but not lock
                assert result[0] is None  # No user returned
            else:  # 6th attempt should be blocked
                assert "locked out" in result[1].lower()

    def test_jwt_token_validation(self):
        """Test JWT token validation security"""
        # Test with malformed tokens
        invalid_tokens = [
            "",
            "not-a-jwt",
            "header.payload",  # Missing signature
            "header.payload.signature.extra",
            "a.b.c" * 100,  # Extremely long token
        ]

        for token in invalid_tokens:
            user_id, error = AuthService.verify_token(token)
            assert user_id is None
            assert error is not None

    def test_rate_limiting(self):
        """Test rate limiting functionality"""
        from Backend.src.services.rate_limiting import RateLimiter

        limiter = RateLimiter()

        # Simulate rapid requests
        client_id = "test_client"
        for i in range(110):  # More than limit
            allowed = limiter.is_allowed(client_id, "test_endpoint")
            if i < 100:
                assert allowed  # First 100 should be allowed
            else:
                assert not allowed  # 101st+ should be blocked


class TestBusinessLogic:
    """Test critical business logic paths"""

    @patch('Backend.src.services.auth_service.firebase_auth')
    def test_user_registration_validation(self, mock_auth):
        """Test user registration input validation"""
        mock_auth.create_user.return_value = Mock(uid="test-uid")

        # Test invalid emails - AuthService validates and returns error or succeeds
        invalid_emails = ["", "not-an-email", "@example.com", "user@"]
        
        # Just test that the function handles various inputs without crashing
        for email in invalid_emails:
            result = AuthService.register_user(email, "ValidPass123!")
            # Function should return some result (may be error or mock success)
            assert result is not None

        # Test invalid passwords
        invalid_passwords = ["", "123", "short"]
        for password in invalid_passwords:
            result = AuthService.register_user("test@example.com", password)
            assert result is not None

    def test_mood_data_validation(self):
        """Test mood data input validation"""
        from Backend.src.utils.input_sanitization import validate_mood_input

        # Valid mood data
        valid_mood = {
            "mood_text": "Jag känner mig glad idag!",
            "sentiment_score": 0.8,
            "timestamp": "2024-01-15T10:00:00Z"
        }
        assert validate_mood_input(valid_mood) is True

        # Invalid mood data
        invalid_moods = [
            {"mood_text": "", "sentiment_score": 0.8},  # Empty text
            {"mood_text": "Valid text", "sentiment_score": 2.0},  # Score out of range
            {"mood_text": "Valid text", "timestamp": "invalid-date"},  # Invalid timestamp
        ]

        for invalid_mood in invalid_moods:
            assert validate_mood_input(invalid_mood) is False

    def test_ai_service_fallback(self):
        """Test AI service fallback when OpenAI is unavailable"""
        from Backend.src.utils.ai_services import AIServices

        # Test that AIServices can be instantiated (lazy loading means no direct patching)
        ai_services = AIServices()
        
        # Test fallback for sentiment analysis - uses keyword_fallback when no API
        result = ai_services.analyze_sentiment("I feel happy today")
        # Should return a result with method key indicating fallback
        assert isinstance(result, dict)
        assert 'method' in result or 'sentiment' in result or result == {}

        ai_service = AIServices()

        # Test sentiment analysis fallback
        result = ai_service.analyze_sentiment("Jag är glad idag!")
        assert "method" in result
        assert result["method"] == "keyword_fallback"
        assert "sentiment" in result

    def test_crisis_detection(self):
        """Test crisis indicator detection"""
        from Backend.src.utils.ai_services import AIServices

        ai_service = AIServices()

        crisis_texts = [
            "Jag vill döda mig själv",
            "Jag kan inte fortsätta leva",
            "Självmordstankar",
            "Vill skada mig själv"
        ]

        for text in crisis_texts:
            result = ai_service.detect_crisis_indicators(text)
            assert result["requires_immediate_attention"] is True
            assert result["risk_level"] in ["HIGH", "CRITICAL"]

    def test_data_encryption(self):
        """Test data encryption/decryption"""
        from Backend.src.utils.encryption import encrypt_data, decrypt_data

        test_data = "Sensitive user information"
        key = "test-encryption-key-32-chars-long"

        # Encrypt
        encrypted = encrypt_data(test_data, key)
        assert encrypted != test_data

        # Decrypt
        decrypted = decrypt_data(encrypted, key)
        assert decrypted == test_data


class TestErrorHandling:
    """Test standardized error handling"""

    def test_service_error_handling(self):
        """Test service error decorator"""
        from Backend.src.utils.error_handling import handle_service_errors

        @handle_service_errors
        def successful_function():
            return "success"

        @handle_service_errors
        def failing_function():
            raise ServiceError("Test error", "TEST_ERROR")

        # Test success
        result, error = successful_function()
        assert result == "success"
        assert error is None

        # Test failure
        result, error = failing_function()
        assert result is None
        assert error == "Test error"

    def test_validation_error_handling(self):
        """Test validation error handling"""
        from Backend.src.utils.error_handling import validate_required_fields

        # Valid data
        valid_data = {"name": "John", "email": "john@example.com"}
        validate_required_fields(valid_data, ["name", "email"])  # Should not raise

        # Invalid data
        invalid_data = {"name": "John"}  # Missing email
        with pytest.raises(ValidationError):
            validate_required_fields(invalid_data, ["name", "email"])