Backend/requirements.txt
# Core Flask Dependencies
Flask==3.0.3
Flask-CORS==3.0.10
Flask-Limiter>=3.0.0
Flask-Babel>=3.0.0
Werkzeug==3.0.1

# JWT Authentication
PyJWT>=2.0.0

# Firebase
firebase-admin>=6.0.1

# Environment & Configuration
python-dotenv==1.0.0

# HTTP & Requests
requests>=2.31.0

# Security & Encryption
bcrypt==4.0.1
pycryptodome>=3.0.0
cryptography>=41.0.0
bleach>=6.0.0

# Caching & Rate Limiting
redis>=4.0.0
hiredis>=2.0.0

# AI & ML - OpenAI
openai>=1.0.0

# AI & ML - Google Cloud
google-cloud-speech>=2.0.0
google-cloud-language>=2.0.0
google-cloud-storage>=2.0.0
google-cloud-aiplatform>=1.0.0

# ML & Data Science
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
joblib>=1.1.0

# Voice & Audio Processing
pyttsx3==2.90
SpeechRecognition==3.8.1
sounddevice==0.4.5
soundfile==0.11.0

# PyAudio requires platform-specific wheels and fails to build on Python 3.12+.
# Install it manually when needed (e.g., via pipwin or prebuilt wheels).

# Natural Language Processing
transformers>=4.0.0
librosa>=0.9.2
torch>=2.0.0

# Payment Processing
stripe>=5.0.0

# Database & Storage
google-cloud-firestore>=2.0.0

# Email Services
sendgrid>=6.0.0
resend==0.8.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
marshmallow>=3.0.0

# CSRF Protection - Temporarily disabled for API testing
# Flask-WTF>=1.0.0

# API Documentation
apispec==6.6.1
flasgger>=0.9.5

# Monitoring & Logging
prometheus-client>=0.14.0
sentry-sdk[flask]>=1.9.0

# Background Tasks
celery>=5.2.0
kombu>=5.2.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.18.0
unittest2>=1.1.0
selenium>=4.0.0
coverage>=6.0.0

# Code Quality
flake8>=4.0.0
black>=22.0.0
pylint>=2.12.0

# Health Data Integration
oauthlib>=3.2.0
requests-oauthlib>=1.3.0

# Utilities
Pillow>=9.0.0
python-dateutil>=2.8.0
pytz>=2021.3

# Two-Factor Authentication
pyotp>=2.6.0
qrcode[pil]>=7.3.1
webauthn>=1.0.0

# Production Server
gunicorn>=20.1.0
gevent>=21.12.0

# Development Tools
ipython>=8.0.0
ipdb>=0.13.0